conventional-commit-linter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conventional-commit-linter"
version = "0.1.0"
description = "Enforces Conventional Commits in Git hooks and CI"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "Arya Sianati"}]
maintainers = [{name = "Arya Sianati"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development",
]
requires-python = ">=3.11"
dependencies = [
    "typer >= 0.9.0",
    "rich >= 13.0.0",
    "pydantic >= 2.0.0",
    "gitpython >= 3.1.40",
    "pyyaml >= 6.0",
]

[project.scripts]
conventional-commit-linter = "conventional_commit_linter.cli:app"

[project.optional-dependencies]
dev = [
    "pytest >= 7.4.0",
    "pytest-cov >= 4.0.0",
    "ruff >= 0.4.0",
]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "T20", "Q", "ARG", "PTH", "ERA", "RSE", "PD", "PGH", "PL", "RET", "RUF", "SLP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "F405"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/conventional_commit_linter"]
